♻️ [Refactoring]: [FE] resolveCloseTarget を providers/AppViewProvider 配下へ移設
AppView 型と同居させるため、`src/hooks/useAppView/` 配下に取り残されていた
resolveCloseTarget と関連テストを `src/providers/AppViewProvider/` 配下へ移し、
import を `./index` 経由に揃える。useAppView/ フォルダは空になるため削除。
