+++ src/providers/AppViewProvider/__tests__/resolveCloseTarget.behavior.test.ts
